requirements.txt
streamlit 
pandas 
Pillow 
reportlab 
requests 
geopy 
qrcode